--- middlewares/auth.js
@@ -1,10 +1,0 @@
-var middlewares = {};
-
-middlewares.isLoggedIn = function(req, res, next){
-    if(req.isAuthenticated()){
-        return next();
-    }
-    res.redirect("/login");
-}
-
-module.exports = middlewares;
