+++ middlewares/auth.middleware.js
@@ -1,0 +1,10 @@
+var middlewares = {};
+
+middlewares.isLoggedIn = function(req, res, next){
+    if(req.isAuthenticated()){
+        return next();
+    }
+    res.redirect("/login");
+}
+
+module.exports = middlewares;
