--- models/user.js
@@ -1,12 +1,0 @@
-var mongoose = require("mongoose");
-var passportLocalMongoose = require("passport-local-mongoose");
-
-var UserSchema = new mongoose.Schema({
-    username:String,
-    password:String
-});
-
-UserSchema.plugin(passportLocalMongoose);
-
-module.exports = mongoose.model("User",UserSchema);
-
