+++ models/user.models.js
@@ -1,0 +1,12 @@
+var mongoose = require("mongoose");
+var passportLocalMongoose = require("passport-local-mongoose");
+
+var UserSchema = new mongoose.Schema({
+    username:String,
+    password:String
+});
+
+UserSchema.plugin(passportLocalMongoose);
+
+module.exports = mongoose.model("User",UserSchema);
+
