+++ views/Authmodule/pages/login.ejs
@@ -1,5 +1,5 @@
-<%- include("partials/header",{page:page ? page:null}) %>
-<%- include("partials/background") %>
+<%- include("./../../partials/header",{page:page ? page:null}) %>
+<%- include("./../../partials/background") %>
 
 <div class="formpage">
 
@@ -16,7 +16,7 @@
     </div>
 
     <div class="logininfo">
-        <form action="/login" method="POST">
+        <form action="/auth/login" method="POST">
             <div class="infodiv">
                 <i class="far fa-user fieldicons"></i>
                 <input class="info" type="text" placeholder="Username">
@@ -36,13 +36,13 @@
             Or Connect Using
         </div>
         <div class="altOptions">
-            <img class="pageicons" src="./img/google.jpg" alt="">
-            <img class="pageicons" src="./img/fb.jpg" alt="">
+            <img class="pageicons" src="/img/google.jpg" alt="">
+            <img class="pageicons" src="/img/fb.jpg" alt="">
         </div>
         <div class="otheroption">
-            Don't have an account? <a class="otherlink" href="/signup">Sign Up</a>
+            Don't have an account? <a class="otherlink" href="/auth/signup">Sign Up</a>
         </div>
     </div>
 </div>
 
-<%- include("partials/footer") %>
+<%- include("./../../partials/footer") %>
