+++ views/partials/header.ejs
@@ -7,11 +7,12 @@
 
     <script src="https://kit.fontawesome.com/2635296004.js" crossorigin="anonymous"></script>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    <link rel="stylesheet" href="./css/login.css">
-    <link rel="stylesheet" href="./css/signup.css"> 
-    <link rel="stylesheet" href="./css/feed.css"> 
+    <link rel="stylesheet" href="/css/login.css">
+    <link rel="stylesheet" href="/css/signup.css"> 
+    <link rel="stylesheet" href="/css/feed.css"> 
 
     <title><%=page%></title>
+    <base href="/">
 </head>
 <body>
